issue-73 Test correction
